+++ energyquantified/api/instances.py
@@ -1,6 +1,5 @@
 from .base import BaseAPI
 
-from ..exceptions import ValidationError
 from ..metadata import CurveType
 from ..parser.metadata import parse_instance_list
 from ..parser.timeseries import parse_timeseries, parse_timeseries_list
@@ -106,7 +105,8 @@ class InstancesAPI(BaseAPI):
             aggregation=None,
             hour_filter=None,
             threshold=None,
-            ensembles=False):
+            ensembles=False,
+            unit=None):
         """
         Load time series instances.
 
@@ -148,13 +148,15 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
-        :param threshold: Allow that many values to be missing within one frame of \
-            *frequency*. Has no effect unless *frequency* is provided, \
-            defaults to 0.
+        :param threshold: Allow that many values to be missing within one \
+            frame of *frequency*. Has no effect unless *frequency* is \
+            provided, defaults to 0.
         :type threshold: int, optional
         :param ensembles: Whether to include ensembles where available,\
             defaults to False
         :type ensembles: bool, optional
+        :param unit: Convert unit of data, defaults to curves unit
+        :type unit: str, optional
         :return: A list :py:class:`energyquantified.data.Timeseries` instances
         :rtype: list
         """
@@ -180,6 +182,7 @@ class InstancesAPI(BaseAPI):
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
             self._add_int(params, "threshold", threshold, min=0)
+        self._add_str(params, "unit", unit)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries_list(response.json())
@@ -194,7 +197,8 @@ class InstancesAPI(BaseAPI):
             aggregation=None,
             hour_filter=None,
             threshold=None,
-            ensembles=False):
+            ensembles=False,
+            unit=None):
         """
         Get the latest time series instance with filtering on `tags` and
         `issued_at_latest`.
@@ -222,13 +226,15 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
-        :param threshold: Allow that many values to be missing within one frame of \
-            *frequency*. Has no effect unless *frequency* is provided, \
+        :param threshold: Allow that many values to be missing within one \
+            frame of *frequency*. Has no effect unless *frequency* is provided,\
             defaults to 0.
         :type threshold: int, optional
         :param ensembles: Whether to include ensembles where available,\
             defaults to False
         :type ensembles: bool, optional
+        :param unit: Convert unit of data, defaults to curves unit
+        :type unit: str, optional
         :return: A time series instance
         :rtype: :py:class:`energyquantified.data.Timeseries`
         """
@@ -249,6 +255,7 @@ class InstancesAPI(BaseAPI):
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
             self._add_int(params, "threshold", threshold, min=0)
+        self._add_str(params, "unit", unit)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries(response.json())
@@ -263,7 +270,8 @@ class InstancesAPI(BaseAPI):
             aggregation=None,
             hour_filter=None,
             threshold=None,
-            ensembles=False):
+            ensembles=False,
+            unit=None):
         """
         Get an instance specified by a `issued` (issue date) and `tag`. The
         default tag is blank and tags are case-insensitive.
@@ -291,13 +299,15 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
-        :param threshold: Allow that many values to be missing within one frame of \
-            *frequency*. Has no effect unless *frequency* is provided, \
+        :param threshold: Allow that many values to be missing within one \
+            frame of *frequency*. Has no effect unless *frequency* is provided,\
             defaults to 0.
         :type threshold: int, optional
         :param ensembles: Whether to include ensembles where available,\
             defaults to False
         :type ensembles: bool, optional
+        :param unit: Convert unit of data, defaults to curves unit
+        :type unit: str, optional
         :return: A time series instance
         :rtype: :py:class:`energyquantified.data.Timeseries`
         """
@@ -318,6 +328,7 @@ class InstancesAPI(BaseAPI):
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
             self._add_int(params, "threshold", threshold, min=0)
+        self._add_str(params, "unit", unit)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries(response.json())
@@ -333,11 +344,13 @@ class InstancesAPI(BaseAPI):
             time_of_day=None,
             after_time_of_day=None,
             before_time_of_day=None,
+            modified_at_latest=None,
             time_zone=None,
             frequency=None,
             aggregation=None,
             hour_filter=None,
-            threshold=None):
+            threshold=None,
+            unit=None):
         """
         Load one instance for each day based on some common criteria, stitch
         them together and return a continuous time series.
@@ -349,8 +362,10 @@ class InstancesAPI(BaseAPI):
 
         You may control the time of the day the instance is issued by setting
         exactly one of the follow parameters: ``time_of_day``,
-        ``after_time_of_day`` or ``before_time_of_day``. These should be set
-        to a time (HH:MM:SS). You can use the :py:class:`datetime.time`.
+        ``after_time_of_day`` or ``before_time_of_day``. Additionally, you may
+        control the time of the day the instance has been modified (or created
+        if modified is null) by setting ``modified-at-latest``. These should be
+        set to a time (HH:MM:SS). You can use the :py:class:`datetime.time`.
 
         This operation works for curves with ``curve_type = INSTANCE`` only.
 
@@ -376,6 +391,9 @@ class InstancesAPI(BaseAPI):
         :param before_time_of_day:  The instance must be issued before this\
             time of day, defaults to None
         :type before_time_of_day: time, str, optional
+        :param modified-at-latest:  The instance must be modified (or created\
+            if modified is null) before this time of day, defaults to None
+        :type modified-at-latest: time, str, optional
         :param time_zone: Set the timezone for the date-times
         :type time_zone: TzInfo, optional
         :param frequency: Set the preferred frequency for aggregations,\
@@ -387,10 +405,12 @@ class InstancesAPI(BaseAPI):
         :param hour_filter: Filters on hours to include (i.e. BASE, PEAK),\
             has no effect unless *frequency* is provided, defaults to BASE
         :type hour_filter: Filter, optional
-        :param threshold: Allow that many values to be missing within one frame of \
-            *frequency*. Has no effect unless *frequency* is provided, \
+        :param threshold: Allow that many values to be missing within one \
+            frame of *frequency*. Has no effect unless *frequency* is provided, \
             defaults to 0.
         :type threshold: int, optional
+        :param unit: Convert unit of data, defaults to curves unit
+        :type unit: str, optional
         :return: A time series
         :rtype: :py:class:`energyquantified.data.Timeseries`
         """
@@ -409,19 +429,14 @@ class InstancesAPI(BaseAPI):
         self._add_time(params, "time-of-day", time_of_day)
         self._add_time(params, "after-time-of-day", after_time_of_day)
         self._add_time(params, "before-time-of-day", before_time_of_day)
+        self._add_time(params, "modified-at-latest", modified_at_latest)
         self._add_time_zone(params, "timezone", time_zone, required=False)
         self._add_frequency(params, "frequency", frequency)
         if "frequency" in params:
             self._add_aggregation(params, "aggregation", aggregation)
             self._add_filter(params, "hour-filter", hour_filter)
             self._add_int(params, "threshold", threshold, min=0)
-        # Additional validation checks
-        if sum(1 if t is not None else 0 for t in
-               (time_of_day, after_time_of_day, before_time_of_day)) > 1:
-            raise ValidationError(reason=(
-                "At most one of the following fields must be set: "
-                "time_of_day, after_time_of_day, before_time_of_day"
-            ))
+        self._add_str(params, "unit", unit)
         # HTTP request
         response = self._get(url, params=params)
         return parse_timeseries(response.json())