+++ CHANGELOG.rst
@@ -2,10 +2,22 @@ Changelog
 =========
 
 dev
-----
-- Add ``unit`` and ``denominator`` attributes to ``Series`` and ``OHLCList``
-- Add support for unit conversion when loading from the API
+---
+
+**Improvements**
+
+- Add ``PeriodInstancesAPI.relative()`` for relative queries
+- Add ``modified-at-latest`` parameter for ``eq.instances.relative()``
 - Implement absolute forecasts for instances ``eq.instances.absolute()``
+
+**Bugfixes**
+
+- Remove unnecessary limitation from ``eq.instances.relative()`` for parameters
+  ``time_of_day``, ``after_time_of_day`` and ``before_time_of_day``
+
+**Breaking changes**
+
+- Removed ``Place.area``. Use ``Place.areas`` instead.
 
 
 0.11